Theros Beyond Death update
The blue changes are straightforward upgrades (reanimator will be
marginally hurt by losing Search for Azcanta).  New Kiora was a
consideration, but I'm a little worried about repetitive gameplay there,
as I think the flicker effect would be very snowball-y.

Bag of Holding was an obvious oversight, and Triskelion is underpowered.

The new Nylea is substantially better than the old, and The Great Henge
is seeing  play in Standard, which tips it over the edge
from "undeniably sweet" to "include".  Beanstalk Giant has a stronger
legacy than Growth Spasm at this point, and feels Green-er.

City of Brass is largely because I found that I have one.

Multicolor was brought closer to being evenly split among the
color combinations.  For a non-gold set, Theros had surprisingly good
options here. The Ashiok swap was partly motivated by the minus ability
being the most-powerful Dinrova Horror variant, partly that 3-mana
Ashiok was in practice a mill card.

Goblin Dark-Dwellers -> Chandra is a pretty straightforward upgrade.
Wrecking Ogre -> Embercleave is, probably, as well, but the sword is now
iconic. Experimental Frenzy has been very good in Standard, but without
the Steam-Kin to back it up it's somewhat underpowered.

Spectral Procession has been on the chopping block for a while, and
Taranika manages to be both an interesting card in its own right and a
pretty neat color-pie-correct take on Geist of Saint Traft.