Noticed that Mana Crypt wasn't in the pool, decided to trim down to one version of each Eldrazi titan
I'd really prefer an Ulamog somewhere in between the nigh-strictly-better Ceaseless Hunger and Infinite Gyre,
because I actively dislike the "exile top 20" on Ceaseless Hunger as it is de facto game-winning, and I think it's
also a pretty hard flavor-fail. On the other hand, the casting cost and cast trigger are strictly better on CH.

I've been of the opinion that Promised End is the correct Emrakul for quite some time, because while Aeons Torn
is more iconic, Promised End is actually castable, and represents a more interesting payoff for deckbuilding.

Kozilek is the one that I'm legitimately on the fence about. Butcher of Truth is currently getting the nod for
largely-nostalgia reasons, and I believe it also received marginally more competitive play.  I like the counter ability
on Great Distortion, and the peak on the draw trigger is certainly higher.  I also appreciate actually requiring
colorless mana in the cast cost, as that was a set theme that is otherwise unrepresented.

Mana Crypt is here to stay, Eldrazi monument is largely provisional.

The other edits are true-up.