Partial Ikoria update
The Khans tap-lands -> Triomes are a clear upgrade.

Hopefully there's a shard-aligned printing in the near future.

I think that Shark Typhoon is a more interesting card for deckbuilding than Talrand, and the pedigree is also better.

Voracious Greatshark vs. Thing in the Ice is closer: the art and flavor of Thing were both phenomenal, but the actual gameplay was somewhat lacking.

I don't think Jaws is going to revolutionize the cube, but it seems like a sweet option for at least a while.

Nissa, Who Shakes the World should probably have been included earlier, but is definitely in now.

While I like the idea of Sylvan Scrying, Nissa is substantially more powerful.

Kogla is pretty sweet, and giving green a ETB fight effect is nice.

Inspired Ultimatum is a much stronger Jeskai carrot than Narset, Enlightened Master.

I think I also prefer it to Narset of the Ancient Way, although it's possible I'll open another Jeskai slot for that.