Vivien, Monster's Advocate is my kinda gal.
Jade Mage is sweet, but the power level is ultimately just not up to the bar.

Dovin Baan is canonically dead, and doesn't have nearly the pedigree of lil Teferi.

I was somewhat torn on which of (Venser, Dovin) deserves the axe (they're both dead).

Venser mostly only blinks things, but:
  A) that's very U/W
  B) his ultimate actually closes out games when it happens

Narset was also a consideration, because both she and lil Teferi are mono-Blue by color pie, or at least could be.