Innistrad: Midnight Hunt update
The prior omission of Ranger Class was an oversight, now corrected.

The new Chandra package is mostly in the same boat: I think there's a
reasonable argument that Acolyte of Flame deserves a slot because of the
"go wide" theme, but Awakened Inferno and Heart of Fire are both notably
powerful planeswalkers.