AFR oversight: Profane Command -> Power Word Kill
Profane Command is a sweet card, and it does have a minor constructed
pedigree.  The biggest problem with it is that while it technically has
4 modes, in practice it's only getting cast if it either A) wins the
game via life loss B) wins the game via Fear C) approximates a Nekrataal
by killing an opposing creature and reanimating one of yours.

With the general stats creep, in practice X needs to be 4+, and this is
just not one of the more powerful plays at CMC 6+

Power Word Kill gets some flavor points.  .