OTJ update
The multi-color overhaul (particularly for GB) reflects a philosophical
shift to wanting gold cards to be more of a draw rather than a reward
for being in the combination. Part of that is just that it's now more
viable: the drastically smaller card pool meant that the best options
were often more in the realm of utility than bomb.

The goblin subtheme in red is largely a thing of the past, and while
Garrison -> Posse is arguably a downgrade, it at least avoids the
awkwardness of the potential expectation of Battlements. I've liked the
depth of play from Mercenary tokens in OTJ.

A few other cards (notably, Tinybones) are definitely under
consideration.