Modern Horizons 3 update
Some of this really ends up being a retroactive MH2 update, but MH3
incarnation reprints catalyzed the inclusions.

Cutting invitational cards is a little unfortunate, but the reality is
that Dark Confidant is simply not very good in Cube.

Psychic Frog vs. Shadowmage Infiltrator is closer, but ultimately I'm
erring on the side of enabling reanimator.

Orcish Bowmasters may prove too situational, but I want to give it a
shot.

Atraxa is pinch-hitting for the Esper 3-color slot because the true
Esper options are basically terrible.

Ghostfire Slice and Breathe Your Last were both likely oversights,
although those are both sufficiently marginal upgrades that it's
unlikely to matter for long.