Foundations / Aetherdrift update
I opened Earthshaker Giant in Mystery Boosters at MagicCon Vegas, so
it's eligible for inclusion.

I'm still on the fence about the merits of the Verge cycle vs. the M10
duals / surveil duals.

Basri was amazing in Aetherdrift sealed, and has some compelling
synergies but I think is ultimately just not quite powerful enough for
this cube.
That lands the Aetherdrift update as (currently) just Draconautics
Engineer, but that is probably the best red 2 drop in the full list.

Fact or Fiction -> Intrude on the Mind is probably the most questionable
edit that's getting included here, but the gist of the thought there is
that it's more interesting due to the flash creature angle.